Refractoring ConfigurationSource.cs for consistency in naming convention
